+++ src/client/cypress/e2e/detailPage/boreholeform.cy.js
@@ -1,9 +1,10 @@
-import { saveLocationForm } from "../helpers/buttonHelpers";
+import { saveWithSaveBar } from "../helpers/buttonHelpers";
 import { clickOnRowWithText, showTableAndWaitForData, sortBy } from "../helpers/dataGridHelpers";
-import { evaluateInput, evaluateSelect, isDisabled, setSelect } from "../helpers/formHelpers";
+import { evaluateInput, evaluateSelect, isDisabled, setInput, setSelect } from "../helpers/formHelpers";
 import {
   createBorehole,
   goToRouteAndAcceptTerms,
+  handlePrompt,
   newEditableBorehole,
   returnToOverview,
   startBoreholeEditing,
@@ -44,7 +45,7 @@ describe("Test for the borehole form.", () => {
     evaluateSelect("qtReferenceElevationId", "20114002");
     evaluateSelect("referenceElevationTypeId", "20117004");
 
-    saveLocationForm();
+    saveWithSaveBar();
     // navigate away and back to check if values are saved
     cy.get('[data-cy="borehole-menu-item"]').click();
     cy.get('[data-cy="location-menu-item"]').click();
@@ -58,19 +59,92 @@ describe("Test for the borehole form.", () => {
 
     // fill all dropdowns on borehole tab
     cy.get('[data-cy="borehole-menu-item"]').click();
-    cy.get('[data-cy="domain-dropdown"]')
-      .should("have.length", 4)
-      .each(el => cy.wrap(el).click().find('[role="option"]').eq(1).click());
+    setSelect("purposeId", 1);
+    setSelect("typeId", 1);
+    setSelect("qtDepthId", 1);
+    setSelect("statusId", 1);
 
-    const boreholeDropdownValues = [];
-    cy.get('[data-cy="domain-dropdown"]')
-      .each(el => {
-        const value = el[0].children[1].firstChild.data;
-        boreholeDropdownValues.push(value);
-      })
-      .then(() => {
-        expect(boreholeDropdownValues).to.deep.eq(["borehole", "geotechnics", "open, no completion", "2"]);
-      });
+    evaluateSelect("purposeId", "22103001");
+    evaluateSelect("typeId", "20101001");
+    evaluateSelect("qtDepthId", "22108001");
+    evaluateSelect("statusId", "22104001");
+
+    saveWithSaveBar();
+
+    // navigate away and back to check if values are saved
+    cy.get('[data-cy="location-menu-item"]').click();
+    cy.get('[data-cy="borehole-menu-item"]').click();
+
+    evaluateSelect("purposeId", "22103001");
+    evaluateSelect("typeId", "20101001");
+    evaluateSelect("qtDepthId", "22108001");
+    evaluateSelect("statusId", "22104001");
+  });
+
+  it("Fills all inputs on borehole tab and saves", () => {
+    createBorehole({ "extended.original_name": "AAA_Ferret", "custom.alternate_name": "AAA_Ferret" }).as("borehole_id");
+    cy.get("@borehole_id").then(id => {
+      goToRouteAndAcceptTerms(`/${id}/borehole`);
+      startBoreholeEditing();
+
+      setSelect("purposeId", 1);
+      setSelect("typeId", 1);
+      setSelect("qtDepthId", 1);
+      setSelect("statusId", 1);
+      setSelect("lithologyTopBedrockId", 1);
+      setSelect("lithostratigraphyId", 1);
+      setSelect("chronostratigraphyId", 1);
+      setSelect("hasGroundwater", 1);
+
+      setInput("totalDepth", 700);
+      setInput("topBedrockFreshMd", 0.60224);
+      setInput("topBedrockWeatheredMd", 78945100);
+      setInput("remarks", "This is a test remark");
+
+      // navigate away is blocked before saving
+      cy.get('[data-cy="location-menu-item"]').click();
+
+      const messageUnsavedChanges = "There are unsaved changes. Do you want to discard all changes?";
+      handlePrompt(messageUnsavedChanges, "cancel");
+
+      saveWithSaveBar();
+      cy.get('[data-cy="location-menu-item"]').click();
+      cy.contains("Boreholes.swissgeol.ch ID");
+    });
+  });
+
+  it("Updates TVD Values when depth values change in boreholeform", () => {
+    createBorehole({ "extended.original_name": "AAA_Ferret", "custom.alternate_name": "AAA_Ferret" }).as("borehole_id");
+    cy.get("@borehole_id").then(id => {
+      goToRouteAndAcceptTerms(`/${id}/borehole`);
+      startBoreholeEditing();
+      setInput("totalDepth", 700);
+      setInput("topBedrockFreshMd", 0.60224);
+      setInput("topBedrockWeatheredMd", 78945100);
+
+      evaluateInput("totalDepth", "700");
+      evaluateInput("topBedrockFreshMd", "0.60224");
+      evaluateInput("topBedrockWeatheredMd", "78'945'100");
+
+      // in display only inputs the label is used for the data-cy instead of the field name
+      evaluateInput("total_depth_tvd", "700");
+      evaluateInput("top_bedrock_fresh_tvd", "0.6");
+      evaluateInput("top_bedrock_weathered_tvd", "78'945'100");
+
+      saveWithSaveBar();
+
+      returnToOverview();
+      showTableAndWaitForData();
+      clickOnRowWithText("AAA_Ferret");
+      cy.get('[data-cy="borehole-menu-item"]').click();
+      evaluateInput("totalDepth", "700");
+      evaluateInput("topBedrockFreshMd", "0.60224");
+      evaluateInput("topBedrockWeatheredMd", "78'945'100");
+
+      evaluateInput("total_depth_tvd", "700");
+      evaluateInput("top_bedrock_fresh_tvd", "0.6");
+      evaluateInput("top_bedrock_weathered_tvd", "78'945'100");
+    });
   });
 
   it("Checks if form values are updated when borehole changes", () => {